# src/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  model.ApiError:
    properties:
      error:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
...
  model.EpisodeCard:
    properties:
      cover_url:
        type: string
      duration_seconds:
        type: integer
      id:
        type: string
      podcast_name:
        type: string
      published_at:
        type: string
      title:
        type: string
    type: object
  model.EpisodeDetail:
    properties:
      cover_url:
        type: string
      duration_seconds:
        type: integer
      id:
        type: string
      podcast_name:
        type: string
      published_at:
        type: string
      title:
        type: string
    type: object
  model.EpisodeDetailResponse:
    properties:
      episode:
        $ref: '#/definitions/model.EpisodeDetail'
    type: object
  model.EpisodeListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/model.EpisodeCard'
        type: array
      next_cursor:
        type: string
      total:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Podcast analysis backend — episodes, topics, transcripts, fact-checks,
    chat.
  title: Audiolens API
  version: "1.0"
paths:
...
  /episodes:
    get:
      description: Cursor-based paginated episode list with optional free-text search.
      parameters:
      - description: Free-text search on title and podcast name
        in: query
        name: q
        type: string
      - description: Cursor from previous response
        in: query
        name: cursor
        type: string
      - description: Page size (default 100, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.EpisodeListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ApiError'
      summary: Episode-Liste
      tags:
      - episodes
  /episodes/{id}:
    get:
      description: Get episode detail for the header area above the tabs.
      parameters:
      - description: Episode ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.EpisodeDetailResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ApiError'
      summary: Episode-Detail
      tags:
      - episodes
...
swagger: "2.0"